Local title, title rating, title ratings

+++ Portfolio-2/DataAccess/TitleRatings.cs
@@ -4,7 +4,7 @@ namespace DataAccess
     {
         public string Tconst { get; set; }
 
-        public int AverageRating { get; set; }
+        public float AverageRating { get; set; }
 
         public int NumVotes { get; set; }
     }
